Clean Version of the Client
