--- exemples/TabSolutionsClient/src/app/settings/settings.component.html
@@ -1,3 +1,0 @@
-<p>
-  settings works!
-</p>
