--- exemples/server/daten.json
@@ -1,17 +1,0 @@
-{
-    "ID1": {
-        "title": "Fairy Tail",
-        "description": "Just another monday morning",
-        "imageURL": "https://www.hevs.ch/img/logo-hes-so-valais.png"
-    },
-    "ID2": {
-        "title": "Sword Art Online",
-        "description": "just another tuesday morning",
-        "imageURL": "https://upload.wikimedia.org/wikipedia/de/thumb/c/ca/Harrypotter-series-logo.svg/1200px-Harrypotter-series-logo.svg.png"
-    },
-    "ID3": {
-        "title": "Fullmetal Alchemist: Brotherhood",
-        "description": "Heftigä schitt",
-        "imageURL": "https://pbs.twimg.com/profile_images/697675802956144640/qy_NOqXT_400x400.jpg"
-    }
-}
