+++ src/TabSolutionsClient/.angular-cli.json
