+++ src/TabSolutionsClient/.editorconfig
