+++ src/TabSolutionsClient/src/app/app.component.html
@@ -5,7 +5,9 @@
 <div id = "nav">
     <a routerLink="/board">Board</a>
     <a routerLink="/settings">Settings</a>
+    <a routerLink="/login">Login</a>
 </div>
+    
 <div id = "content">
 <router-outlet></router-outlet>
 </div>
