+++ src/TabSolutionsClient/src/app/app.component.ts
@@ -9,7 +9,7 @@ import { Component, OnInit } from '@angular/core';
    
 })
 export class AppComponent implements OnInit {
-  title = 'app';
+  title = 'TabSolutions';
   results = '';
 
   constructor(){
