+++ src/TabSolutionsClient/src/app/board/board.component.css
@@ -14,3 +14,7 @@
     background-color: sandybrown; 
     padding: 2%,
 }
+#settings
+{
+    background-color: blue; 
+}
