+++ src/TabSolutionsClient/src/app/board/board.component.html
@@ -4,7 +4,7 @@
   <div id = "title">{{ currentBoard.title }}</div>
 
   <div id = "description">{{currentBoard.description}}</div>
-
+  
   <div id = "image"><img src = "{{ currentBoard.imageURL }}" width="50%"/></div>
 
 </div>
