+++ src/TabSolutionsClient/src/app/settings/settings.component.html
@@ -1,0 +1,14 @@
+<form> 
+  <h3>Kreuzen Sie die gewünschten Zutaten an:</h3> 
+  <fieldset>
+    <ul>
+      <li> 
+        <label>
+          <input type="checkbox" name="zutat" value="salami"/>
+          {{currentBoard.title}}
+        </label>
+      </li>
+    </ul> 
+  </fieldset> 
+</form>
+
