+++ src/TabSolutionsClient/src/app/settings/settings.component.spec.ts
