+++ src/TabSolutionsClient/src/tsconfig.app.json
