+++ src/TabSolutionsClient/src/tsconfig.spec.json
